+++ library/NekMeshUtils/CADSystem/CADCurve.h
@@ -77,12 +77,13 @@ public:
      *
      * @return Array of two entries, min and max parametric coordinate.
      */
-    virtual Array<OneD, NekDouble> GetBounds() = 0;
+    NEKMESHUTILS_EXPORT virtual Array<OneD, NekDouble> GetBounds() = 0;
 
     /**
      * @brief Returns the minimum and maximum parametric coords t of the curve.
      */
-    virtual void GetBounds(NekDouble &tmin, NekDouble &tmax) = 0;
+    NEKMESHUTILS_EXPORT virtual void GetBounds(
+        NekDouble &tmin, NekDouble &tmax) = 0;
 
     /**
      * @brief Calculates the arclength between the two paremetric points \p ti
@@ -92,7 +93,8 @@ public:
      * @param tf Second parametric coordinate.
      * @return Arc length between \p ti and \p tf.
      */
-    virtual NekDouble Length(NekDouble ti, NekDouble tf) = 0;
+    NEKMESHUTILS_EXPORT virtual NekDouble Length(
+        NekDouble ti, NekDouble tf) = 0;
 
     /**
      * @brief Gets the location (x,y,z) in an array out of the curve at
@@ -101,7 +103,7 @@ public:
      * @param t Parametric coordinate
      * @return Array of x,y,z
      */
-    virtual Array<OneD, NekDouble> P(NekDouble t) = 0;
+    NEKMESHUTILS_EXPORT virtual Array<OneD, NekDouble> P(NekDouble t) = 0;
 
     /**
      * @brief Gets the location (x,y,z) in an array out of the curve at
@@ -109,17 +111,18 @@ public:
      *
      * @param t Parametric coordinate
      */
-    virtual void P(NekDouble t, NekDouble &x, NekDouble &y, NekDouble &z) = 0;
+    NEKMESHUTILS_EXPORT virtual void P(
+        NekDouble t, NekDouble &x, NekDouble &y, NekDouble &z) = 0;
 
     /**
      * @brief Gets the second derivatives at t
      */
-    virtual Array<OneD, NekDouble> D2(NekDouble t) = 0;
+    NEKMESHUTILS_EXPORT virtual Array<OneD, NekDouble> D2(NekDouble t) = 0;
 
     /**
      * @brief Calculates the radius of curvature of the curve at point t
      */
-    virtual NekDouble Curvature(NekDouble t) = 0;
+    NEKMESHUTILS_EXPORT virtual NekDouble Curvature(NekDouble t) = 0;
 
     /**
      * @brief Calculates the parametric coordinate and arclength location
@@ -130,14 +133,14 @@ public:
      *
      * @todo This really needs improving for accuracy.
      */
-    virtual NekDouble tAtArcLength(NekDouble s) = 0;
+    NEKMESHUTILS_EXPORT virtual NekDouble tAtArcLength(NekDouble s) = 0;
 
     /**
      * @brief Gets the start and end of the curve.
      *
      * @return Array with 6 entries of endpoints x1,y1,z1,x2,y2,z2.
      */
-    virtual Array<OneD, NekDouble> GetMinMax() = 0;
+    NEKMESHUTILS_EXPORT virtual Array<OneD, NekDouble> GetMinMax() = 0;
 
     /**
      * @brief set the ids of the surfaces either side of the curve
@@ -186,25 +189,26 @@ public:
      * @brief locates a point in the parametric space. returns the
      * distance to the point and passes t by reference and updates it
      */
-    virtual NekDouble loct(Array<OneD, NekDouble> xyz, NekDouble &t) = 0;
+    NEKMESHUTILS_EXPORT virtual NekDouble loct(
+        Array<OneD, NekDouble> xyz, NekDouble &t) = 0;
 
     /**
      * @brief Returns the orientation of the curve with respect to a given
      * surface by id surf
      */
-    CADOrientation::Orientation GetOrienationWRT(int surf);
+    NEKMESHUTILS_EXPORT CADOrientation::Orientation GetOrienationWRT(int surf);
 
     /**
      * @brief Returns the normal to the curve which is orientate with respect
      * to the surface surf
      */
-    Array<OneD, NekDouble> NormalWRT(NekDouble t, int surf);
+    NEKMESHUTILS_EXPORT Array<OneD, NekDouble> NormalWRT(NekDouble t, int surf);
 
     /**
      * @brief Returns the normal to a curve, it will always point in the concave
      * direction
      */
-    virtual Array<OneD, NekDouble> N(NekDouble t) = 0;
+    NEKMESHUTILS_EXPORT virtual Array<OneD, NekDouble> N(NekDouble t) = 0;
 
 protected:
     /// Length of edge