+++ CHANGELOG.md
@@ -9,6 +9,10 @@ v5.1.0
 
 **CardiacEPSolver**
 - Added additional parameter sets to Fenton-Karma model (!1119)
+
+**NekMesh**
+- Improved boundary layer splitting and output to CADfix (!938)
+
 v5.0.1
 ------
 **Library**